fix: query countEgressoByEmpresas

+++ backend/meuegresso/src/main/java/labes/facomp/ufpa/br/meuegresso/dto/grafico/EmpresaGraficoDTO.java
@@ -18,6 +18,6 @@ public class EmpresaGraficoDTO {
     
     private String empresa;
 
-    private Integer quantidade;
+    private Long quantidade;
 
 }
